User can save an invoice [Finishes #77762264] -Add frontend validations -Add backend model validations -Store mileage_rate as a decimal -Use $location service to redirect to Show Invoice view

+++ db/migrate/20140916115940_create_service_reports.rb
@@ -10,7 +10,7 @@ class CreateServiceReports < ActiveRecord::Migration
       t.integer :travel
       t.integer :travel_rate
       t.integer :mileage
-      t.integer :mileage_rate            
+      t.decimal :mileage_rate            
       t.decimal :sales_tax
       t.timestamps
     end
